+++ src/MockXMLHttpRequestUpload.ts
@@ -1,0 +1,4 @@
+import MockXMLHttpRequestEventTarget from './MockXMLHttpRequestEventTarget';
+
+export default class MockXMLHttpRequestUpload extends MockXMLHttpRequestEventTarget
+  implements XMLHttpRequestUpload {}